==> your-code/1. web_scraping.ipynb <==
# Web Scraping

## Libraries

# main
import pandas as pd
import numpy as np

# web scraping
from bs4 import BeautifulSoup
import requests

## Web Scraping "Billboard 100"

### "Soup" Preparation

On this few lines of code, I will prepare the "soup" for getting the songs from the **Billboard 100**

url = "https://www.billboard.com/charts/hot-100"

response = requests.get(url)

response

soup = BeautifulSoup(response.content, "html.parser")

### Scraping songs

First of all, let's test the code to obtain the name of the first song. Once the result returns the same output as the one from the web, I will loop through it to obtaing the rest of them.

soup.select("span.chart-element__information__song.text--truncate.color--primary")[0].get_text()

song = []

for s in soup.select("span.chart-element__information__song.text--truncate.color--primary"):
    song.append(s.get_text())

# Check that the lenght is 100, the amount of songs it should have
len(song)

### Scraping artists

As with the songs, first of all let's find the line of code that returs the first artist of the list. Once we know it's the correct one, then we will loop through it like the step before.

soup.select("span.chart-element__information__artist.text--truncate.color--secondary")[0].get_text()

artist = []

for a in soup.select("span.chart-element__information__artist.text--truncate.color--secondary"):
    artist.append(a.get_text())

# Check that the lenght is 100, the amount of artists it should have
len(artist)

## Creating DataFrame

Once we have the **Song Name** and the **Artist**, we can create a `DataFrame` for better visualization and being able to store it as a *csv* for the next part of the project.

billboard_hot100 = pd.DataFrame({"song":song,
                         "artist(s)":artist
                         })

billboard_hot100

# Storing it as a CSV
billboard_hot100.to_csv("../data/billboard100.csv", index=False)
