==> your-code/2. spotify_song_features.ipynb <==
# Spotify Song Features

## Libraries

# main
import pandas as pd
import numpy as np

# spotify API
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

## Spotify API

### Initialize API

The **Spotify API** has to be initilized with the `client_id` and with the `client_secret`. As the `client_secret` contains sensitive information, I have created a local file, read it with Python and then loop through it, asigning each line of the document to the correspondant variable.

credentials = []

with open("spotify_credentials.txt", "r") as spotify:
    for line in spotify:
        line = line.rstrip("\n")
        credentials.append(line)

#Initialize SpotiPy with user credentias
spotify = spotipy.Spotify(client_credentials_manager=SpotifyClientCredentials(
    client_id = credentials[0],
    client_secret = credentials[1]))

### Obtaining Playlist

For this initial project, we will find a big Spotify Playlist (*10K Songs*) for ensuring that our sample is going to have different features, so the output predicted could be as much accurate as possible. 

To obtain all the inforamation, I will build a function able to retrieve all the information available of that playlist.

def get_playlist_tracks(username,playlist_id):
    """
    Input: Username of the API initialized and ID of the playlist we want to pass
    Output: The playlist itself
    """
    results = spotify.user_playlist_tracks(username,playlist_id)
    tracks = results['items']
    while results['next']:
        results = spotify.next(results)
        tracks.extend(results['items'])
    return tracks

playlist = get_playlist_tracks("spotify", "spotify:playlist:1G8IpkZKobrIlXcVPoSIuf")

Once we have the `playlist`, the next thing we have to do is retain only the content that offers us value for that project. In our case, it will be **`song_names`**, **`artist_names`**, **`song_features`** and **`duration_ms`**. 

Mention that, for obtain the `song_features` first of all we will have to retrieve the **`song_URI`** and, after that, another function will pull that information.

def playlist_info(playlist):
    """
    Input: The playlist stored from the previous function
    Output: The parts from the playlist that are useful for us, in this case songs, artists and features
    """
    song_names = []
    artist_names = []
    duration = []
    song_URI = []
    features = []
    i = 0
    
    for result in playlist:
        try:
            if result["is_local"] == False:
                i += 1
                song_names.append(result["track"]["name"])
                artist_names.append(result["track"]["artists"][0]["name"])
                duration.append(result["track"]["duration_ms"])
                song_URI.append(result["track"]["uri"])
                
                # Once Spotify reach 100, the speed for loading URI slows down, that's why I'm setting 100 as threshold
                # This will append lists with 100 lenght
                if i == 100:
                    i = 0
                    features.append(spotify.audio_features(song_URI))
                    song_URI = []
                    
        except TypeError:
            pass
    
    features.append(spotify.audio_features(song_URI)) 
    
    # Creating a list that matches the same lenght as the other variables
    correct_feature = [f for feature in features for f in feature]
    
    return song_names, artist_names, duration, correct_feature

songs, artists, duration, feature = playlist_info(playlist)

As we do have the `uri` from each song, now it's time to obtaing the final result, the **`song_features`**. After looking into the different features, I will select the ones that I consider it have more impact for predicting a song based on the user input.

def features_df_creation(info_features):
    """
    Input: List of URIs from the previous function
    Output: DF containing the song, artist, duration and the song_features that we chose
    """
    acousticness = []
    danceability = []
    energy = []
    instrumentalness = []
    liveness = []
    loudness = []
    speechiness = []
    valence = []
    tempo = []
    
    for feature in info_features:
        acousticness.append(feature["acousticness"])
        danceability.append(feature["danceability"])
        energy.append(feature["energy"])
        instrumentalness.append(feature["instrumentalness"])
        liveness.append(feature["liveness"])
        loudness.append(feature["loudness"])
        speechiness.append(feature["speechiness"])
        valence.append(feature["valence"])
        tempo.append(feature["tempo"])
        
    
    df = pd.DataFrame({"song":songs,
                       "artist":artists,
                       "duration":duration,
                       "danceability":danceability,
                       "energy":energy,
                       "acousticness":acousticness,
                       "instrumentalness":instrumentalness,
                       "liveness":liveness,
                       "loudness":loudness,
                       "speechiness":speechiness,
                       "valence":valence,
                       "tempo":tempo
                      })
    
    return df

complete_list_song_artist_features = features_df_creation(feature)

complete_list_song_artist_features.head()

# Storing it as a CSV
complete_list_song_artist_features.to_csv("../data/spotify_playlist.csv", index=False)
